backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
httpx
pydantic
langchain
langchain-community
langchain-core
langchain-text-splitters
chromadb
openai
moviepy
Pillow
python-multipart
tiktoken
youtube-transcript-api
trafilatura
beautifulsoup4
gspread
google-auth
